GSdx: Adjust default best render on first use/setup.
For intel set the default render to OpenGL if igpu supports feature
level above 11.1. According to the information I didn't see an igpu that
has support for 11.2 but not 12 so we can use the detection to select GL
by default. Skylake has all the necessary features to properly support
pcsx2.